feat: Add "Wrap with RepaintBoundary" command
This commit adds a new command, "Wrap with RepaintBoundary", to the Flutter development extension. The command allows users to easily wrap any widget with a `RepaintBoundary` widget, which isolates the repaint process of the wrapped widget. This can improve performance in complex UIs.

Note: This commit message follows the established convention of using a "feat" prefix for new feature additions.